+++ .gitignore
@@ -22,3 +22,4 @@ storybook-static
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+reko
