+++ .husky/pre-commit
@@ -3,4 +3,4 @@
 
 npm run lint 
 npm run prettify
-npm run build-storybook
+npm run build:components
